Doc formatting, Complex64 float and helper fixes
Improve XML doc generation and type linking, change Complex64 to use single precision, and tidy helper code.

- OnnxNodeGenerator: reworked operator/parameter/attribute comments (bold headings, include domain/version, link allowed types with <see/>, improved default null display) and updated MarkdownHelper to handle inline code blocks, paragraph splitting and escaping more robustly.
- Complex64: change Real/Imaginary and constructor to float (single precision) to match ONNX expectations.
- OnnxHelper: use MemoryMarshal.AsBytes unqualified, prefer var in loops, adjust packing logic for complex float data (write floats directly) and minor bit-conversion adjustments.

These changes improve generated documentation clarity, ensure correct float precision for complex numbers, and clean up helper implementations.

+++ src/Onnxify/Data/Complex64.cs
@@ -2,10 +2,10 @@
 
 public readonly struct Complex64
 {
-    public double Real { get; }
-    public double Imaginary { get; }
+    public float Real { get; }
+    public float Imaginary { get; }
 
-    public Complex64(double real, double imaginary)
+    public Complex64(float real, float imaginary)
     {
         Real = real;
         Imaginary = imaginary;
